Change 64 to 32 bit where possible, and refactor to reduce mem pressure
By moving a lot of the loadStars functionality to a new method, we can free up
memory in the meantime and greatly reduce the full overhead by ~20%.

+++ fgcm/fgcmParameters.py
@@ -288,7 +288,7 @@ class FgcmParameters(object):
         self.parLnPwvQuadratic = np.zeros(self.campaignNights.size, dtype=np.float64)
 
         # parameters with per-epoch values
-        self.parSuperStarFlat = np.zeros((self.nEpochs,self.nLUTFilter,self.nCCD,self.superStarNPar),dtype=np.float64)
+        self.parSuperStarFlat = np.zeros((self.nEpochs,self.nLUTFilter,self.nCCD,self.superStarNPar),dtype=np.float32)
         # The first term should be 1.0 with new flux units
         self.parSuperStarFlat[:, :, :, 0] = 1.0
